nop-repeat fix
Repeat now re-uses the existing DOM parent element instead of replacing
the to-be-repeated item with a DIV. This allows us to e.g.
nop-repeat a "tr" inside a "table". Not sure if this will break
anything though.
After creating the repeated elements, attach event handlers to these
new elements specifically so that changes are propagated automatically.